# ROX-19888: better sorting of listening endpoints