--- a/central/processlisteningonport/store/postgres/full_store.go
+++ b/central/processlisteningonport/store/postgres/full_store.go
@@ -2,6 +2,7 @@ package postgres
 
 import (
 	"context"
+	"sort"
 	"time"
 
 	"github.com/gogo/protobuf/proto"
@@ -79,6 +80,23 @@ func (s *fullStoreImpl) retryableGetPLOP(
 	}
 
 	return results, rows.Err()
+}
+
+func plopComparison(plop1 *storage.ProcessListeningOnPort, plop2 *storage.ProcessListeningOnPort) bool {
+        if plop1.PodId != plop2.PodId {
+                return plop1.PodId < plop2.PodId
+        }
+        if plop1.Signal.ExecFilePath != plop2.Signal.ExecFilePath {
+                return plop1.Signal.ExecFilePath < plop2.Signal.ExecFilePath
+        }
+        if plop1.Endpoint.Port != plop2.Endpoint.Port {
+                return plop1.Endpoint.Port < plop2.Endpoint.Port
+        }
+        return plop1.Endpoint.Protocol < plop2.Endpoint.Protocol
+}
+
+func sortPlops(plops []*storage.ProcessListeningOnPort) {
+	sort.Slice(plops, func(i, j int) bool { return plopComparison(plops[i], plops[j]) })
 }
 
 // Manual converting of raw data from SQL query to ProcessListeningOnPort (not
@@ -168,6 +186,8 @@ func (s *fullStoreImpl) readRows(
 
 		plops = append(plops, plop)
 	}
+
+	sortPlops(plops)
 
 	log.Debugf("Read returned %+v plops", len(plops))
 	return plops, nil