--- a/central/processlisteningonport/store/postgres/full_store.go
+++ b/central/processlisteningonport/store/postgres/full_store.go
@@ -83,16 +83,16 @@ func (s *fullStoreImpl) retryableGetPLOP(
 }
 
 func plopComparison(plop1 *storage.ProcessListeningOnPort, plop2 *storage.ProcessListeningOnPort) bool {
-        if plop1.PodId != plop2.PodId {
-                return plop1.PodId < plop2.PodId
-        }
-        if plop1.Signal.ExecFilePath != plop2.Signal.ExecFilePath {
-                return plop1.Signal.ExecFilePath < plop2.Signal.ExecFilePath
-        }
-        if plop1.Endpoint.Port != plop2.Endpoint.Port {
-                return plop1.Endpoint.Port < plop2.Endpoint.Port
-        }
-        return plop1.Endpoint.Protocol < plop2.Endpoint.Protocol
+	if plop1.PodId != plop2.PodId {
+		return plop1.PodId < plop2.PodId
+	}
+	if plop1.Signal.ExecFilePath != plop2.Signal.ExecFilePath {
+		return plop1.Signal.ExecFilePath < plop2.Signal.ExecFilePath
+	}
+	if plop1.Endpoint.Port != plop2.Endpoint.Port {
+		return plop1.Endpoint.Port < plop2.Endpoint.Port
+	}
+	return plop1.Endpoint.Protocol < plop2.Endpoint.Protocol
 }
 
 func sortPlops(plops []*storage.ProcessListeningOnPort) {