--- a/central/processlisteningonport/store/postgres/sort_test.go
+++ b/central/processlisteningonport/store/postgres/sort_test.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/stackrox/rox/generated/storage"
+	"github.com/stackrox/rox/pkg/fixtures/fixtureconsts"
 	"github.com/stretchr/testify/suite"
 )
 
@@ -75,7 +76,7 @@ func makeRandomPlops(nport int, nprocess int, npod int) []*storage.ProcessListen
 }
 
 
-func (s *SortSuite) TestSort1000() {
+func (suite *SortSuite) TestSort1000() {
 	nport := 10
 	nprocess := 10
 	npod := 10
@@ -90,7 +91,7 @@ func (s *SortSuite) TestSort1000() {
 
 }
 
-func (s *SortSuite) TestSort8000() {
+func (suite *SortSuite) TestSort8000() {
 	nport := 20
 	nprocess := 20
 	npod := 20
@@ -105,7 +106,7 @@ func (s *SortSuite) TestSort8000() {
 
 }
 
-func (s *SortSuite) TestSort125000() {
+func (suite *SortSuite) TestSort125000() {
 	nport := 50
 	nprocess := 50
 	npod := 50
@@ -120,7 +121,7 @@ func (s *SortSuite) TestSort125000() {
 
 }
 
-func (s *SortSuite) TestSort1000000() {
+func (suite *SortSuite) TestSort1000000() {
 	nport := 100
 	nprocess := 100
 	npod := 100
@@ -130,11 +131,99 @@ func (s *SortSuite) TestSort1000000() {
 	sortPlops(plops)
 	duration := time.Since(startTime)
 
-	// var m runtime.MemStats
-	// runtime.ReadMemStats(&m)
-	// fmt.Printf("Memory usage: %d bytes\n", m.Alloc)
-
 
 	fmt.Printf("Sorting %d took %s\n", len(plops), duration)
 
+}
+
+// func plopComparison(plop1 *storage.ProcessListeningOnPort, plop2 *storage.ProcessListeningOnPort) bool {
+// 	if plop1.PodId != plop2.PodId {
+// 		return plop1.PodId < plop2.PodId
+// 	}
+// 	if plop1.Signal.ExecFilePath != plop2.Signal.ExecFilePath {
+// 		return plop1.Signal.ExecFilePath < plop2.Signal.ExecFilePath
+// 	}
+// 	if plop1.Endpoint.Port != plop2.Endpoint.Port {
+// 		return plop1.Endpoint.Port < plop2.Endpoint.Port
+// 	}
+// 	return plop1.Endpoint.Protocol < plop2.Endpoint.Protocol
+// }
+
+func (suite *SortSuite) TestSortVarious() {
+
+	execFilePath1 := "app"
+	execFilePath2 := "zap"
+
+	plop1 := storage.ProcessListeningOnPort{
+		Endpoint: &storage.ProcessListeningOnPort_Endpoint{
+			Port:     80,
+			Protocol: storage.L4Protocol_L4_PROTOCOL_TCP,
+		},
+		DeploymentId:  fixtureconsts.Deployment1,
+		PodId:         fixtureconsts.PodName1,
+		PodUid:        fixtureconsts.PodUID1,
+		Signal: &storage.ProcessSignal{
+			ExecFilePath: execFilePath1,
+		},
+	}
+
+	plop2 := storage.ProcessListeningOnPort{
+		Endpoint: &storage.ProcessListeningOnPort_Endpoint{
+			Port:     1234,
+			Protocol: storage.L4Protocol_L4_PROTOCOL_TCP,
+		},
+		DeploymentId:  fixtureconsts.Deployment1,
+		PodId:         fixtureconsts.PodName1,
+		PodUid:        fixtureconsts.PodUID1,
+		Signal: &storage.ProcessSignal{
+			ExecFilePath: execFilePath1,
+		},
+	}
+
+	plop3 := storage.ProcessListeningOnPort{
+		Endpoint: &storage.ProcessListeningOnPort_Endpoint{
+			Port:     1234,
+			Protocol: storage.L4Protocol_L4_PROTOCOL_UDP,
+		},
+		DeploymentId:  fixtureconsts.Deployment1,
+		PodId:         fixtureconsts.PodName1,
+		PodUid:        fixtureconsts.PodUID1,
+		Signal: &storage.ProcessSignal{
+			ExecFilePath: execFilePath1,
+		},
+	}
+
+	plop4 := storage.ProcessListeningOnPort{
+		Endpoint: &storage.ProcessListeningOnPort_Endpoint{
+			Port:     1234,
+			Protocol: storage.L4Protocol_L4_PROTOCOL_TCP,
+		},
+		DeploymentId:  fixtureconsts.Deployment1,
+		PodId:         fixtureconsts.PodName1,
+		PodUid:        fixtureconsts.PodUID1,
+		Signal: &storage.ProcessSignal{
+			ExecFilePath: execFilePath2,
+		},
+	}
+
+	plop5 := storage.ProcessListeningOnPort{
+		Endpoint: &storage.ProcessListeningOnPort_Endpoint{
+			Port:     1234,
+			Protocol: storage.L4Protocol_L4_PROTOCOL_TCP,
+		},
+		DeploymentId:  fixtureconsts.Deployment1,
+		PodId:         fixtureconsts.PodName2,
+		PodUid:        fixtureconsts.PodUID2,
+		Signal: &storage.ProcessSignal{
+			ExecFilePath: execFilePath1,
+		},
+	}
+
+	plops := []*storage.ProcessListeningOnPort{&plop3, &plop5, &plop1, &plop2, &plop4}
+
+	sortPlops(plops)
+
+	expectedSortedPlops := []*storage.ProcessListeningOnPort{&plop1, &plop2, &plop3, &plop4, &plop5}
+
+	suite.Equal(expectedSortedPlops, plops)
 }