--- a/central/processlisteningonport/store/postgres/sort_test.go
+++ b/central/processlisteningonport/store/postgres/sort_test.go
@@ -1,10 +1,7 @@
 package postgres
 
 import (
-	"fmt"
-	"math/rand"
 	"testing"
-	"time"
 
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/fixtures/fixtureconsts"
@@ -17,118 +14,6 @@ type SortSuite struct {
 
 func TestSortSuite(t *testing.T) {
 	suite.Run(t, new(SortSuite))
-}
-
-func makeRandomString(length int) string {
-	var charset = []byte("asdfqwert")
-	randomString := make([]byte, length)
-	for i := range randomString {
-		randomString[i] = charset[rand.Intn(len(charset))]
-	}
-	return string(randomString)
-}
-
-func makeRandomPlops(nport int, nprocess int, npod int) []*storage.ProcessListeningOnPort {
-	deploymentID := makeRandomString(10)
-	count := 0
-
-	plops := make([]*storage.ProcessListeningOnPort, 2*nport*nprocess*npod)
-	for podIdx := 0; podIdx < npod; podIdx++ {
-		podID := makeRandomString(10)
-		podUID := makeRandomString(10)
-		for processIdx := 0; processIdx < nprocess; processIdx++ {
-			execFilePath := makeRandomString(10)
-			for port := 0; port < nport; port++ {
-
-				plopTCP := &storage.ProcessListeningOnPort{
-					Endpoint: &storage.ProcessListeningOnPort_Endpoint{
-						Port:     uint32(port),
-						Protocol: storage.L4Protocol_L4_PROTOCOL_TCP,
-					},
-					DeploymentId: deploymentID,
-					PodId:        podID,
-					PodUid:       podUID,
-					Signal: &storage.ProcessSignal{
-						ExecFilePath: execFilePath,
-					},
-				}
-				plopUDP := &storage.ProcessListeningOnPort{
-					Endpoint: &storage.ProcessListeningOnPort_Endpoint{
-						Port:     uint32(port),
-						Protocol: storage.L4Protocol_L4_PROTOCOL_UDP,
-					},
-					DeploymentId: deploymentID,
-					PodId:        podID,
-					PodUid:       podUID,
-					Signal: &storage.ProcessSignal{
-						ExecFilePath: execFilePath,
-					},
-				}
-				plops[count] = plopTCP
-				count++
-				plops[count] = plopUDP
-				count++
-			}
-		}
-	}
-
-	return plops
-}
-
-func (suite *SortSuite) TestSort1000() {
-	nport := 10
-	nprocess := 10
-	npod := 10
-	plops := makeRandomPlops(nport, nprocess, npod)
-
-	startTime := time.Now()
-	sortPlops(plops)
-	duration := time.Since(startTime)
-
-	fmt.Printf("Sorting %d took %s\n", len(plops), duration)
-
-}
-
-func (suite *SortSuite) TestSort8000() {
-	nport := 20
-	nprocess := 20
-	npod := 20
-	plops := makeRandomPlops(nport, nprocess, npod)
-
-	startTime := time.Now()
-	sortPlops(plops)
-	duration := time.Since(startTime)
-
-	fmt.Printf("Sorting %d took %s\n", len(plops), duration)
-
-}
-
-func (suite *SortSuite) TestSort125000() {
-	nport := 50
-	nprocess := 50
-	npod := 50
-	plops := makeRandomPlops(nport, nprocess, npod)
-
-	startTime := time.Now()
-	sortPlops(plops)
-	duration := time.Since(startTime)
-
-	fmt.Printf("Sorting %d took %s\n", len(plops), duration)
-
-}
-
-func (suite *SortSuite) TestSort1000000() {
-	nport := 100
-	nprocess := 100
-	npod := 100
-	plops := makeRandomPlops(nport, nprocess, npod)
-
-	startTime := time.Now()
-	sortPlops(plops)
-	duration := time.Since(startTime)
-
-	fmt.Printf("Sorting %d took %s\n", len(plops), duration)
-
 }
 
 func (suite *SortSuite) TestSortVarious() {