--- a/qa-tests-backend/src/test/groovy/ProcessesListeningOnPortsTest.groovy
+++ b/qa-tests-backend/src/test/groovy/ProcessesListeningOnPortsTest.groovy
@@ -112,7 +112,6 @@ class ProcessesListeningOnPortsTest extends BaseSpecification {
         assert list.size() == 2
 
         def endpoint1 = list[0]
-        //def endpoint1 = list.find { it.endpoint.port == 80 }
 
         assert endpoint1
         assert endpoint1.deploymentId
@@ -124,7 +123,6 @@ class ProcessesListeningOnPortsTest extends BaseSpecification {
         assert endpoint1.signal.args == "-d -d -v TCP-LISTEN:80,fork STDOUT"
 
         def endpoint2 = list[1]
-        //def endpoint2 = list.find { it.endpoint.port == 8080 }
 
         assert endpoint2
         assert endpoint2.deploymentId