## Description

Sorts API responses by podID, executable file path, port, and protocol.

## Checklist
- [x] Investigated and inspected CI test results
- [x] Unit test and regression tests added
- [x] Checked the ordering in the UI
- [x] Bench marking
- [x] Load testing
~~- [ ] Evaluated and added CHANGELOG entry if required~~
~~- [ ] Determined and documented upgrade steps~~
~~- [ ] Documented user facing changes (create PR based on [openshift/openshift-docs](https://github.com/openshift/openshift-docs) and merge into [rhacs-docs](https://github.com/openshift/openshift-docs/tree/rhacs-docs))~~

If any of these don't apply, please comment below.

## Testing Performed

### Here I tell how I validated my change

Checking the UI

![Screenshot from 2023-10-02 18-30-22](https://github.com/stackrox/stackrox/assets/24723000/58b0bac9-d3e4-4289-98f2-6c85a1d902ef)

The UI has the desired ordering.

### Benchmarking

The time used by the sorting function was measured on another branch, https://github.com/stackrox/stackrox/pull/8299

```
git checkout jv-ROX-19888-better-sorting-of-listening-endpoints-testing
cd central/processlisteningonport/store/postgres
go test
Sorting 2000 took 218.638µs
Sorting 2000000 took 554.243997ms
Sorting 250000 took 55.002619ms
Sorting 16000 took 2.102102ms
```

Not sure how much extra memory was used.

### Benchmarking with db

```
git checkout jv-ROX-19888-better-sorting-of-listening-endpoints-testing
cd central/processlisteningonport/datastore

go test -tags=sql_integration -v
```

Without sorting 
```
=== RUN   TestPLOPDataStore/TestSort1000000
Fetching 2000000 plops 2.419662924s took
```

With sorting
```
=== RUN   TestPLOPDataStore/TestSort1000000
Fetching 2000000 plops 3.804959441s took
```
In production the number of fetched plops will be orders of magnitude less than this.

### Load testing

The testing version of this branch was checked out. It had some improvements for the load generation.

```
git checkout jv-ROX-19888-better-sorting-of-listening-endpoints-testing
```

The following script was used generate load for testing
```
#!/usr/bin/env bash
set -eou pipefail

cluster_name=$1
tag=${2:-}

export CLUSTER_NAME="$cluster_name"
export MAIN_IMAGE_TAG=$tag

if [[ -z "$tag" ]]; then
	tag="$(make tag)"
fi
echo "tag= $tag"

build=true

if [[ $build == "true" ]]; then
    printf 'yes\n'  | /home/jvirtane/go/src/github.com/stackrox/workflow/scripts/runtime/teardown.sh &
    make image
    make central-db-image
    make docker-build-roxctl-image
    docker tag quay.io/stackrox-io/main:$tag quay.io/jvirtane/main:$tag
    docker push quay.io/jvirtane/main:$tag
    docker tag quay.io/stackrox-io/central-db:$tag quay.io/jvirtane/central-db:$tag
    docker push quay.io/jvirtane/central-db:$tag
    docker tag quay.io/stackrox-io/roxctl:$tag quay.io/jvirtane/roxctl:$tag
    docker push quay.io/jvirtane/roxctl:$tag
else
    printf 'yes\n' | /home/jvirtane/go/src/github.com/stackrox/workflow/scripts/runtime/teardown.sh
fi

infractl get $CLUSTER_NAME --json | jq '.Connect' -r | bash

export USE_MIDSTREAM_IMAGES=false

export MAIN_IMAGE_REPO=quay.io/jvirtane/main
export CENTRAL_DB_IMAGE_REPO=quay.io/jvirtane/central-db
export ROXCTL_IMAGE_REPO=quay.io/jvirtane/roxctl

export API_ENDPOINT=localhost:8000

export STORAGE=pvc # Backing storage
export STORAGE_CLASS=faster # Runs on an SSD type
export STORAGE_SIZE=100 # 100G
export MONITORING_SUPPORT=true # Runs monitoring
export LOAD_BALANCER=lb

./deploy/k8s/central.sh # Launches central

kubectl -n stackrox port-forward deploy/central 8000:8443 > /dev/null 2>&1 &

sleep 20

export ROX_ADMIN_USERNAME=admin

password="$(cat deploy/k8s/central-deploy/password)"

export ROX_ADMIN_PASSWORD="$password"

./deploy/k8s/sensor.sh

kubectl -n stackrox set env deploy/sensor MUTEX_WATCHDOG_TIMEOUT_SEC=0
kubectl -n stackrox set env deploy/sensor ROX_FAKE_KUBERNETES_WORKLOAD=long-running

kubectl -n stackrox set env deploy/central MUTEX_WATCHDOG_TIMEOUT_SECS=0

./scale/launch_workload.sh sort-plop

```

You may need to make manual changes to push to your repository instead.

The workload for the test was the following
```
$ cat scale/workloads/sort-plop.yaml
deploymentWorkload:
- deploymentType: Deployment
  lifecycleDuration: 72h
  numDeployments: 1
  numLifecycles: 0
  podWorkload:
    containerWorkload:
      numImages: 0
    lifecycleDuration: 72h
    numContainers: 30
    numPods: 50
    processWorkload:
      alertRate: 0.001
      processInterval: 1s
  updateInterval: 10m0s
networkWorkload:
  batchSize: 5000
  flowInterval: 30s
nodeWorkload:
  numNodes: 1
rbacWorkload:
  numBindings: 100
  numRoles: 100
  numServiceAccounts: 100
```
The number of deployments was set to 1 so that when the API was called all of the listening endpoints would be sorted and central would be under maximum stress. The number of pods and containers was increased compensate for the low number of deployments. lifecycleDuration was set to a long time, so that the pods and their associated listening endpoints would not be deleted.

The image repository for sensor was set manually after running the script.

```
central_active=# select count(*) from listening_endpoints ;
-[ RECORD 1 ]-
count | 199883

central_active=# select count(*) from process_indicators ;
-[ RECORD 1 ]
count | 9993
```

```
central_active=# select distinct deploymentid from listening_endpoints ;
-[ RECORD 1 ]+-------------------------------------
deploymentid | 80afe7c2-0c93-46b6-a7d6-886ea521116e
```

This confirms that there is only one deployment.

Once this number of records were in the database, sensor was killed so that the number of records would not change during the tests.

```
ks delete deployment sensor
```

The following script was run generate the test file

```
$ cat MakeK6.sh 
#!/usr/bin/env bash
set -eoux pipefail

outfile=$1

port=8000

logmein localhost:$port &> token_file.txt

token="$(cat token_file.txt | sed 's|.*token=||' | sed 's|&type.*||')"

password="$(cat ./deploy/k8s/central-deploy/password)"

curl -sSkf -u "admin:$password" -o /dev/null -w '%{redirect_url}' "https://localhost:$port/sso/providers/basic/4df1b98c-24ed-4073-a9ad-356aec6bb62d/challenge?micro_ts=0"

deployments="$(curl --location --request GET "https://localhost:$port/v1/deployments" -k -H "Authorization: Bearer $token")"


deployment="$(echo "$deployments" | jq -r .deployments[0].id)"

ndeployment="$(echo $deployments | jq '.deployments | length')"

cp k6-script-prefix.js $outfile

for ((i = 0; i < ndeployment; i = i + 1)); do
    deployment="$(echo "$deployments" | jq -r .deployments[$i].id)"

    echo "deployment= $deployment"
    echo "  postman[Request]({" >> $outfile
    echo "    name: \"https://localhost:$port/v1/listening_endpoints/deployment/$deployment\"," >> $outfile
    echo "    id: \"$deployment\"," >> $outfile
    echo "    method: \"GET\"," >> $outfile
    echo "    address:" >> $outfile
    echo "      \"https://localhost:$port/v1/listening_endpoints/deployment/$deployment\"," >> $outfile
    echo "    auth(config, Var) {" >> $outfile
    echo "      config.headers.Authorization =" >> $outfile
    echo "        \"Bearer $token\";" >> $outfile
    echo "    }," >> $outfile
    echo "  });" >> $outfile
    echo "" >> $outfile
done


echo "}" >> $outfile
```

```
$ cat k6-script-prefix.js
import "/home/jvirtane/.nvm/versions/node/v20.7.0/lib/node_modules/postman-to-k6/lib/shim/core.js";

export let options = { maxRedirects: 4, insecureSkipTLSVerify: true };

const Request = Symbol.for("request");
postman[Symbol.for("initial")]({
  options,
});

export default function () {
```

The command to generate the test file was 

```
/MakeK6.sh k6-sort-plop-delete-2.js
```


The following script was run to perform load testing and monitor CPU and memory usage during the test.

```
$ cat K6AndTopPod.sh 
#!/usr/bin/env bash
set -eou pipefail

suffix=$1

./TopPod.sh &> top_pod_${suffix}.txt & 
sleep 5
k6 run /home/jvirtane/go/src/github.com/stackrox/stackrox/k6-sort-plop-delete-2.js --vus 5 --iterations 10 --out csv=k6-sort-plop-delete.csv &> k6_load_testing_${suffix}.txt
```

With TopPod.sh having the following
```
#!/usr/bin/env bash
set -eou pipefail

while true; do
	kubectl -n stackrox top pod
	sleep 1
done
```

#### Results with sorting

```
./K6AndTopPod.sh with_sort
```

```
          /\      |‾‾| /‾‾/   /‾‾/
     /\  /  \     |  |/  /   /  /
    /  \/    \    |     (   /   ‾‾\
   /          \   |  |\  \ |  (‾)  |
  / __________ \  |__| \__\ \_____/ .io

  execution: local
     script: /home/jvirtane/go/src/github.com/stackrox/stackrox/k6-sort-plop-delete-2.js
     output: csv (k6-sort-plop-delete.csv)

  scenarios: (100.00%) 1 scenario, 5 max VUs, 10m30s max duration (incl. graceful stop):
           * default: 10 iterations shared among 5 VUs (maxDuration: 10m0s, gracefulStop: 30s)


     data_received..................: 696 MB 13 MB/s
     data_sent......................: 1.1 MB 20 kB/s
     http_req_blocked...............: avg=103.11ms min=264ns   med=98.75ms  max=208.83ms p(90)=208.75ms p(95)=208.79ms
     http_req_connecting............: avg=71.02µs  min=0s      med=66.51µs  max=152.72µs p(90)=147.35µs p(95)=150.03µs
     http_req_duration..............: avg=25.77s   min=9.82s   med=25.71s   max=34.32s   p(90)=33.93s   p(95)=34.13s
       { expected_response:true }...: avg=25.77s   min=9.82s   med=25.71s   max=34.32s   p(90)=33.93s   p(95)=34.13s
     http_req_failed................: 0.00%  ✓ 0       ✗ 10
     http_req_receiving.............: avg=16.63s   min=5.89s   med=16.61s   max=22.87s   p(90)=22.85s   p(95)=22.86s
     http_req_sending...............: avg=254.35µs min=44.01µs med=244.67µs max=490.96µs p(90)=490.46µs p(95)=490.71µs
     http_req_tls_handshaking.......: avg=102.48ms min=0s      med=98.11ms  max=207.56ms p(90)=207.49ms p(95)=207.52ms
     http_req_waiting...............: avg=9.13s    min=3.93s   med=9.23s    max=11.67s   p(90)=11.41s   p(95)=11.54s
     http_reqs......................: 10     0.18141/s
     iteration_duration.............: avg=25.87s   min=9.82s   med=25.81s   max=34.53s   p(90)=34.14s   p(95)=34.34s
     iterations.....................: 10     0.18141/s
     vus............................: 2      min=2     max=5
     vus_max........................: 5      min=5     max=5


running (00m55.1s), 0/5 VUs, 10 complete and 0 interrupted iterations
default ✓ [ 100% ] 5 VUs  00m55.1s/10m0s  10/10 shared iters
```

#### Results without sorting

```
git checkout jv-ROX-19888-better-sorting-of-listening-endpoints-testing-without-sort

make image
```

Push the image if needed. Change the image for central manually.
Run the load test with monitoring.

```
./K6AndTopPod.sh without_sort
```

```
          /\      |‾‾| /‾‾/   /‾‾/
     /\  /  \     |  |/  /   /  /
    /  \/    \    |     (   /   ‾‾\
   /          \   |  |\  \ |  (‾)  |
  / __________ \  |__| \__\ \_____/ .io

  execution: local
     script: /home/jvirtane/go/src/github.com/stackrox/stackrox/k6-sort-plop-delete-2.js
     output: csv (k6-sort-plop-delete.csv)

  scenarios: (100.00%) 1 scenario, 5 max VUs, 10m30s max duration (incl. graceful stop):
           * default: 10 iterations shared among 5 VUs (maxDuration: 10m0s, gracefulStop: 30s)


     data_received..................: 696 MB 13 MB/s
     data_sent......................: 1.1 MB 20 kB/s
     http_req_blocked...............: avg=100.76ms min=242ns   med=97.22ms max=204.25ms p(90)=204.2ms  p(95)=204.22ms
     http_req_connecting............: avg=45.05µs  min=0s      med=41.52µs max=94.68µs  p(90)=93.74µs  p(95)=94.21µs
     http_req_duration..............: avg=26.16s   min=23.8s   med=26.36s  max=27.5s    p(90)=27.47s   p(95)=27.49s
       { expected_response:true }...: avg=26.16s   min=23.8s   med=26.36s  max=27.5s    p(90)=27.47s   p(95)=27.49s
     http_req_failed................: 0.00%  ✓ 0        ✗ 10
     http_req_receiving.............: avg=16.81s   min=16.24s  med=16.66s  max=17.88s   p(90)=17.46s   p(95)=17.67s
     http_req_sending...............: avg=88.09µs  min=45.84µs med=85.49µs max=168.12µs p(90)=123.9µs  p(95)=146.01µs
     http_req_tls_handshaking.......: avg=100.24ms min=0s      med=96.67ms max=203.16ms p(90)=203.14ms p(95)=203.15ms
     http_req_waiting...............: avg=9.34s    min=7.07s   med=9.39s   max=11.19s   p(90)=11.05s   p(95)=11.12s
     http_reqs......................: 10     0.187838/s
     iteration_duration.............: avg=26.27s   min=23.8s   med=26.47s  max=27.71s   p(90)=27.67s   p(95)=27.69s
     iterations.....................: 10     0.187838/s
     vus............................: 3      min=3      max=5
     vus_max........................: 5      min=5      max=5


running (00m53.2s), 0/5 VUs, 10 complete and 0 interrupted iterations
default ✓ [ 100% ] 5 VUs  00m53.2s/10m0s  10/10 shared iters
```

#### Comparison of load testing with and without sorting

![Screenshot from 2023-10-25 16-47-43](https://github.com/stackrox/stackrox/assets/24723000/15b0f793-f502-4c54-8578-edb8bcae0b3a)

Here the initial baseline memory usage before the API call, is subtracted from each point. The maximum memory usage is approximately the same. However the peak in the memory usage lasts much longer when there is sorting, than without. The duration of high memory usage is longer when there is sorting.

![Screenshot from 2023-10-25 16-58-52](https://github.com/stackrox/stackrox/assets/24723000/df28ebe2-bac5-40b7-80de-ba7dc6734c52)

Again the peak CPU memory usage is very similar for the two cases, but the duration of high memory usage is greater for sorting. The CPU usage is high until only about ~70 second mark. This suggests that the high memory usage after 70 seconds is not because the listening endpoints are still being operated on, but because they are being held in memory until the API response is done being sent, which takes a long time as there are ~200,000 listening endpoints.

Looking at the K6 output. There doesn't seem to be much difference. The most important thing is that in both cases all of the requests succeeded. 

The next most important things are the p(90) and p(95) for http_req_waiting.

```
--------------------------------------
|             | Sorting | No Sorting |
--------------------------------------
|p(90) (s)    | 11.41   | 11.05      |
--------------------------------------
|p(95) (s)    | 11.54   | 11.12      |
--------------------------------------
```

These are very long times for an API response, but keep in mind that ~200,000 records are processed in this case. There is no significant difference here between sorting and not sorting.

#### Load testing one single API call

The above load testing performed 10 queries to the API. The load testing was done with just one API call to measure the memory used when doing just one API call.

![Screenshot from 2023-10-26 14-31-05](https://github.com/stackrox/stackrox/assets/24723000/fc24d52a-388a-47fc-983d-a18362a3a9a5)

With one API call the memory used increased by 1 Gi as opposed to 2 Gi when doing the load testing with 10 API calls.

In this case there where 194,939 listening_endpoints and 25,225 listening_endpoints.

```
central_active=# select count(*) from listening_endpoints ;
-[ RECORD 1 ]-
count | 194939

central_active=# select count(*) from process_indicators ;
-[ RECORD 1 ]
count | 25225
``` 

### Reminder for reviewers

In addition to reviewing code here, reviewers **must** also review testing and request further testing in case the
performed one does not seem sufficient. As a reviewer, you must not approve the change until you understand the
performed testing and you are satisfied with it.