@erthalion 2023-10-04
Why does sorting have to be implemented in memory on the app side, and not via `order by` on the db side?