@JoukoVirtanen 2023-10-04
> Why does sorting have to be implemented in memory on the app side, and not via `order by` on the db side?

podID and executable file path are used for the sort and they are in the serialized column.