@erthalion 2023-10-05
> > Why does sorting have to be implemented in memory on the app side, and not via `order by` on the db side?
> 
> podID and executable file path are used for the sort and they are in the serialized column.

What is the upper limit, how much data could be sorted in memory without affecting the instance of central (e.g. due to high memory consumption or such)?