@JoukoVirtanen 2023-10-17
> > > > Why does sorting have to be implemented in memory on the app side, and not via `order by` on the db side?
> > > 
> > > 
> > > podID and executable file path are used for the sort and they are in the serialized column.
> > 
> > 
> > What is the upper limit, how much data could be sorted in memory without affecting the instance of central (e.g. due to high memory consumption or such)?
> 
> I should double check, but I think sorting is in place and only requires O(1) extra memory. The time taken is n*log(n) so that should not be a concern either.

I did some bench marking and added the results above.